+++ packages/react-native/android/src/newarch/HotUpdaterModule.kt
@@ -206,7 +206,7 @@ class HotUpdaterModule internal constructor(
         return impl.clearCrashHistory()
     }
 
-    override fun getBaseURL(): String {
+    override fun getBaseURL(): String? {
         val impl = getInstance()
         return impl.getBaseURL()
     }