+++ plugins/rock/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hot-updater/rnef
 
+## 0.30.2
+
+### Patch Changes
+
+- @hot-updater/cli-tools@0.30.2
+- @hot-updater/plugin-core@0.30.2
+
+## 0.30.1
+
+### Patch Changes
+
+- @hot-updater/cli-tools@0.30.1
+- @hot-updater/plugin-core@0.30.1
+
 ## 0.30.0
 
 ### Minor Changes